pyproject.toml
[tool.poetry]
name = "pythontemplate"
version = "0.1.0"
description = ""
authors = ["Dillon <dillonb07dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
pre-commit = "^3.5.0"
sourcery = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
